[Implémentation] : gestion du test de posabilité et envoie de message au cas où il n'y a pas assez de tuile

+++ Carcassheim_unity/Assets/system/joueur.cs
@@ -1,11 +1,11 @@
 public class joueur {
-    private int _idjoueur;
+    private ulong _idjoueur;
     public int _tricheJoueur;
     public int _noteJoueur;
     public int _pointsJoueur;
     public int _abbe;
     public int _meeples;
-    public void initJoueur(int id){
+    public void initJoueur(ulong id){
         _idjoueur=id;
         _tricheJoueur=0;    //initialisé à 0, incremente si triche, à 2 déconnecter joueur
         _noteJoueur=3;      //3 tours à rater avant déconnecter le joueur
@@ -13,7 +13,7 @@ public class joueur {
         _meeples=5;
         _pointsJoueur=0;
     }
-    public int getIdJoueur(){
+    public ulong getIdJoueur(){
         return _idjoueur;
     }
     public void ajoutPoints(int pointsAAjouter){     //Ajouter des points au joueur
